# Conflicts:
#	api/src/org/apache/cloudstack/api/ApiConstants.java
#	api/src/org/apache/cloudstack/api/command/user/template/CopyTemplateCmd.java
#	api/src/org/apache/cloudstack/api/command/user/template/RegisterTemplateCmd.java
#	engine/orchestration/src/com/cloud/vm/VirtualMachineManagerImpl.java
#	plugins/hypervisors/baremetal/src/com/cloud/baremetal/manager/BareMetalTemplateAdapter.java
#	plugins/hypervisors/kvm/src/com/cloud/hypervisor/kvm/resource/wrapper/LibvirtStopCommandWrapper.java
#	plugins/hypervisors/xenserver/src/com/cloud/hypervisor/xenserver/resource/CitrixResourceBase.java
#	server/src/com/cloud/storage/snapshot/SnapshotManager.java
#	server/src/com/cloud/storage/snapshot/SnapshotManagerImpl.java
#	server/src/com/cloud/template/TemplateManagerImpl.java
#	server/test/org/apache/cloudstack/networkoffering/ChildTestConfiguration.java

+++ engine/storage/volume/src/org/apache/cloudstack/storage/volume/VolumeServiceImpl.java
@@ -415,9 +415,27 @@ public class VolumeServiceImpl implements VolumeService {
                     s_logger.info("Volume " + vo.getId() + " is not referred anywhere, remove it from volumes table");
                     volDao.remove(vo.getId());
                 }
+
                 SnapshotDataStoreVO snapStoreVo = _snapshotStoreDao.findByVolume(vo.getId(), DataStoreRole.Primary);
-                if(snapStoreVo != null){
-                    _snapshotStoreDao.remove(snapStoreVo.getId());
+
+                if (snapStoreVo != null) {
+                    long storagePoolId = snapStoreVo.getDataStoreId();
+                    StoragePoolVO storagePoolVO = storagePoolDao.findById(storagePoolId);
+
+                    if (storagePoolVO.isManaged()) {
+                        DataStore primaryDataStore = dataStoreMgr.getPrimaryDataStore(storagePoolId);
+                        Map<String, String> mapCapabilities = primaryDataStore.getDriver().getCapabilities();
+
+                        String value = mapCapabilities.get(DataStoreCapabilities.STORAGE_SYSTEM_SNAPSHOT.toString());
+                        Boolean supportsStorageSystemSnapshots = new Boolean(value);
+
+                        if (!supportsStorageSystemSnapshots) {
+                            _snapshotStoreDao.remove(snapStoreVo.getId());
+                        }
+                    }
+                    else {
+                        _snapshotStoreDao.remove(snapStoreVo.getId());
+                    }
                 }
             } else {
                 vo.processEvent(Event.OperationFailed);
